Renaming util to lib
